Track settings.php
If it had been started from the beginning with a template file and then
an untracked actual settings file, this could work and save trouble.
But it is not compatible with other forks/branches so best to leave all
settings in settings.php and then keep that in version control. It just
means dealing with merge conflicts for every test, but that is ok.